src/requirements.txt
art
ascii_magic
luma.core
luma.oled
opencv-python
picamera
pillow
requests
vidgear[core]
